+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/viewmodel/EmulateViewModel.kt
@@ -19,13 +19,13 @@ import com.flipperdevices.core.ktx.android.vibrateCompat
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
-import com.flipperdevices.keyscreen.api.EmulateHelper
 import com.flipperdevices.keyscreen.api.EmulateProgress
+import com.flipperdevices.keyscreen.api.emulate.AlreadyOpenedAppException
+import com.flipperdevices.keyscreen.api.emulate.EmulateHelper
 import com.flipperdevices.keyscreen.emulate.model.DisableButtonReason
 import com.flipperdevices.keyscreen.emulate.model.EmulateButtonState
 import com.flipperdevices.keyscreen.emulate.model.LoadingState
 import com.flipperdevices.keyscreen.emulate.tasks.CloseEmulateAppTaskHolder
-import com.flipperdevices.keyscreen.emulate.viewmodel.helpers.AlreadyOpenedAppException
 import com.flipperdevices.protobuf.app.Application
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers