+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [REFACTOR] Migrate widgets to WorkManager
 - [Feature] Add mfkey32 attack in debug/options
 - [Feature] Add HUB tab
+- [Feature] Wrap more errors with widget
 - [Feature] Add error screen on mfkey32
 - [Feature] Not found if key not found in mfkey32
 - [Feature] Fix format for user keys